backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
python-telegram-bot